requirements.txt
pandas>=2.2.0
numpy>=1.26.0
geopandas>=0.14.0
shapely>=2.0.0
pyarrow>=15.0.0
rapidfuzz>=3.9.0

streamlit>=1.34.0
folium>=0.16.0
streamlit-folium>=0.20.0
plotly>=5.22.0

sentence-transformers>=2.7.0
pinecone>=5.0.0
langchain>=0.2.0
langgraph>=0.1.0
langchain-community>=0.2.0
langchain-groq>=0.1.5
python-dotenv>=1.0.1

scikit-learn>=1.5.0
jupyter>=1.0.0
ipykernel>=6.29.0
